feat(account): 添加获取收藏电视节目的功能
- 新增 getFavoriteTv 方法，用于获取用户收藏的电视节目
- 添加相关单元测试用例，验证功能正确性- 更新实体类 FoveriteResult，将 voteAverage 字段类型从 Long 改为 Double
- 删除无用的 Main.kt 文件

+++ src/main/kotlin/entity/account/FoveriteResult.kt
@@ -32,7 +32,7 @@ data class Result (
     val title: String,
     val video: Boolean,
     @SerializedName("vote_average")
-    val voteAverage: Long,
+    val voteAverage: Double,
     @SerializedName("vote_count")
     val voteCount: Long
 )
